+++ fluent-syntax/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod ast;
 pub mod parser;
-pub mod parser2;
 pub mod unicode;
